Re-adding building part1- Spawning and moving
Changes:

Can now place and move buildings (had to use hit.point to get the world position of where the raycast hit vs hit.transform.position which gives the position of the object hit. So if for example the floor is hit by the ray then the building goes to {0,0,0} which is incorrect).

Set scrap and solar base to mesh colliders (now a cylinder collision box instead of box or capsule).

Changed methods to not even run unless enabled (if statements in Update() that will call method if enabled vs going to method and checking if enabled).

Moved interaction raycast to Movement() (doesn't need to be in FixedUpdate()).

Thoughts/Ideas:

Should have a list in world that keeps tract of the number of power devices to they can be given number (solar_panel1).