Swapped to Universal Render Pipeline + fixed issue with swapping buildings
Changes:

Created copy of project but uses URP (Universal Render Pipeline) vs default renderer. *Note- URP is supported on engine versions that are 2019.4.8 LTS and above (can't go back to 2018.4.26 if continue to use URP).

Fixed building code so that hitting different key will swap the building vs not (result in error as certain structure is expected but wasn't what was placed).

Generated lighting for scene.