Save, Load and UI
Changes:

Changed Resource.cs script to MinableResource.cs.

Added empty object called "UI" to player and placed "main" and "inventory" UI's under it.
Moved the inventory buttons to "inventory" UI.

Can now Save and Load (WIP- so far it saves the solarCount, generatorCount and player position + rotation).
*Note: Player rotation doesn't load properly (not right angle for camera) and loading multiple times somehow messes with the controls (WASD get mixed up, not sure how).*

Save and Load is part of the World script (made no sense to be in the player script).