Fixed saving issue with rotations
Changes:

Now saves player rotation properly (playerCamera's y had to be set to the players y otherwise it would be facing a different direction).

Pretty sure this is because the camera bases it's rotation on the parent because the camera's rotation becomes 0 yet it was set to the parents Y rotation in the script. Perhaps it subtracted the rotation?